% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocHDSS.R
\name{filterHDSS}
\alias{filterHDSS}
\title{Filter out valid observations in an HDSS data set}
\usage{
filterHDSS(x)
}
\arguments{
\item{x}{the pre-processed HDSS data frame to filter}
}
\value{
A list with three entries:
\itemize{
  \item{report}{a data frame with one row per filter, describing the filtering
                step and listing the number of events removed by it}
  \item{keep}{the record numbers (\code{RecNr}) of the events to be kept
              after filtering}
  \item{remove}{a list of vectors of record numbers, one vector per filter
                step, identifying events that are dropped in this step}
}
}
\description{
Runs a number of filters on a pre-processed data set to identify questionable
events (before birth, in the future etc.)
}
\seealso{
\code{\link{contenTests}}
}

