% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patternTrajectory.R
\name{getPatterns}
\alias{firstEvents}
\alias{getPatterns}
\alias{lastEvents}
\title{Extract patterns of events over time}
\usage{
getPatterns(x, sep = "->")

firstEvents(x)

lastEvents(x)
}
\arguments{
\item{x}{A pre-processed HDSS data set}

\item{sep}{A character expression that is used to separate events in the pattern}
}
\value{
Either a character vector of event sequences, one for each individual,
or a character of first/last events (by event date), one for each individual
}
\description{
Given a pre-processed HDSS data set, return the sequence of events for
each individual subject, or the first/last events respectively
}
\seealso{
\code{\link{preprocHDSS}}
}

