% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patternTrajectory.R
\name{getTransitions}
\alias{getTransitions}
\title{Extract transitions between events}
\usage{
getTransitions(x)
}
\arguments{
\item{x}{A pre-processed HDSS data set}
}
\value{
A list with two numeric matrices: rows and columns are indexed by the
event codes, with rows indicating the from-state and columns the to-states.
The first list element is called \code{counts}, and contains the number of
transitions from one state to the next; the second element is called
\code{deltaT} and contains the total time spent between the from- and the
to-states.
}
\description{
Given a pre-processed HDSS data set, return matrices describing transitions
between events, both counts and aggregated times between events
}
\seealso{
\code{\link{preprocHDSS}}
}

