% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{INDEPTH_Centres}
\alias{INDEPTH_Centres}
\title{INDEPTH HDSS Sites}
\format{A data frame with three variables.}
\description{
An incomplete list of centres that are part of INDEPTH: the project code,
country and site name.
}
\details{
\itemize{
   \item Code. The five-character code for the side used by INDEPTH.
   \item Country. The name of the country where the site is located.
   \item ISOcode. The three-digit ISO code for the country.
   \item Site. The name of the site.
}
}
\references{
http://www.indepth-ishare.org/index.php/catalog/central
}

