% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{INDEPTH_eventCodes}
\alias{INDEPTH_eventCodes}
\title{Event codes for HDSS data file}
\format{A character vector}
\source{
Osman Sankoh and Peter Byass. The INDEPTH Network: filling vital gaps in global epidemiology. Int J Epi 2012; 41:579-588, Table 2
}
\description{
The three-letter codes for different events that are recorded during surveillance
}

