% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{INDEPTH_EventSeq}
\alias{INDEPTH_EventSeq}
\title{Permissible event sequences for HDSS data}
\format{A logical matrix with as many rows/columns as events in HDSS data}
\description{
A logical matrix indicating which events can logically follow each other
in an individual's event history; from-events are indexed by row, to-events
by column.
}

