% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R.HDSS.R
\docType{package}
\name{R.HDSS}
\alias{R.HDSS}
\alias{R.HDSS-package}
\title{Process HDSS demographic data from the INDEPTH project}
\description{
The INDEPTH project collects data in numerous health and demographic
surveillance system (HDSS) field sites in Africa, Asia and Oceania. This
package provides methods for processing the data files distributed through
the INDEPTH website.
}
\details{
Note that while this package has been written to facilitate an existing
collaboration within the INDEPTH project, it should not be considered part of
INDEPTH. It is an independent effort to make available tools for processing
and analysing the data distributed through the project.
}
\references{
Osman Sankoh and Peter Byass. The INDEPTH Network: filling vital gaps in global epidemiology. Int J Epi 2012; 41:579-588
}

