% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recTests.R
\name{recTest-class}
\alias{print.recTest}
\alias{recTest}
\alias{recTest-class}
\alias{summary.recTest}
\title{Output from a record test}
\usage{
recTest(Index, Desc)

\method{print}{recTest}(x, ...)

\method{summary}{recTest}(x, ...)
}
\arguments{
\item{Index}{logical vector of test results}

\item{Desc}{character expression describing the test}

\item{x}{an object of class \code{recTest}}
}
\value{
An object of class \code{recTest} (invisibly for the print-method)
}
\description{
A simple class that wraps the outpur from a record test like e.g.
\code{recTest_Missing} and comes with a print/summary method.
}
\details{
Internally, this is a list with two entries: a logical matrix named
  \code{Index} with as many rows as records and as many columsn as tests, and
  a character vector \code{Desc} with one entry per test. The constructor
  \code{recTest} takes just one test result, but multiple tests can be
  combined using \code{addRecTest}.
}
\seealso{
\code{\link{recTest_Missing}}
}

