% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recTests.R
\name{recTest-functions}
\alias{addRecTest}
\alias{compressRecTest}
\alias{filterOnRecTest}
\alias{recTest-functions}
\title{Working with record test results}
\usage{
addRecTest(x, y)

compressRecTest(x)

filterOnRecTest(data, rt, failed = FALSE)
}
\arguments{
\item{x,y,rt}{Objects of class \code{recTest}}

\item{data}{the data frame for which the record tests were performed}
}
\value{
For \code{addRecTest}, a new \code{recTest} object that combines the
  results from both tests. \cr\cr
  For \code{compressRecTest}, a reduced \code{recTest} object that only
  retains tests that failed at least once.\cr\cr
  For \code{filterOnRecTest}, a subset of the original data frame passed in:
  if \code{failed=FALSE}, the subset of records that passed all tests;
  otherwise the subset of records that failed at least one test, augmented
  by a new column \code{FailedTest} that describes which test(s) each record
  failed.
}
\description{
Functions to combine, compress and apply record test results.
}

