% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recTests.R
\name{recTests-generic}
\alias{recTest_CorrectCodes}
\alias{recTest_InRange}
\alias{recTest_LessOrEqual}
\alias{recTest_Missing}
\alias{recTests-generic}
\title{Tools for checking records in a data frame}
\usage{
recTest_Missing(x, v)

recTest_CorrectCodes(x, v, cc)

recTest_InRange(x, v, lower, upper)

recTest_LessOrEqual(x, v1, v2)
}
\arguments{
\item{x}{the data frame where to check}

\item{v}{a character vector giving full variable names}

\item{cc}{a vector of correct (permissible) codes for variable \code{v}}

\item{lower,upper}{smallest/largest permissible value for variable \code{v}}

\item{v1,v2}{variable names}
}
\value{
An object of class \code{recTest}.
}
\description{
These functions check for records in a data frame that some required
property.
}
\details{
These are tools for checking with records in a data frame have a
   required property (not missing, specified values etc.) Record test of this
   type are usually run after running basic tests on variables as well as
   generic pre-processing to adjuts data formats. The output allows both
   a top-level summary of how many records pass a series of record tests, as
   well as identification of the records that fail one or more of the
   required tests.\cr\cr
   The data frame is the first argument, to permit simple use of the
   pipe-operator.
}
\seealso{
\code{\link{dplyr::'\%>\%'}} \code{\link{varTests-generic}}
}

