% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recTests-HDSS.R
\name{recTests-HDSS}
\alias{coreRecordTests}
\alias{recTests-HDSS}
\alias{vaRecordTests}
\title{Check pre-processed HDSS data}
\usage{
coreRecordTests(x, maxAge = 105, studyPeriod, observationLag = 0.5)

vaRecordTests(x)
}
\arguments{
\item{x}{The data frame to be validated}

\item{maxAge}{A maximum age for checking the valid range of birth dates}

\item{studyPeriod}{A vector of length two, either of dates indicating the
beginning and end of the study period, or of characters expressing dates
that can be passed to `as.Date`. If not specified, the stuyd period is
inferred from the data, see Details.}

\item{observationLag}{A grace period (in years) after the end of the study
period: observation dates are still accepted as prima facie valid if the fall
not more than \code{observationLag} years after the end of study.}
}
\value{
Each test returns a list with two entries:
  \itemize{
    \item{\code{flag}}{ a logical flag hat indicates whether all tests were passsed or not}
    \item{\code{text}}{ a data frame with one column that contains the error messages from the failed tests; empty if all tests were passed}
  }
}
\description{
These functions test and report which records of a pre-processed HDSS data
have valid and consistent values
}
\details{
These functions report on the content of a pre-processed HDSS data
  set (as opposed to \code{coreVariableTests} and \code{vaVariableTests},
  which check the formal requirements for a raw HDSS data set). \cr\cr
  \code{coreRecordTests} checks for valid codes, missing values introduced by
  pre-processing, proper sequence of events, reasonable ranges of dates and
  ages, proper counting of events and female mothers. This is done for the
  same set of core variables as in \code{coreVariableTests}. \cr
  If no study period is specified by the user, it is inferred from the data
  as follows: start of study is the earliest non-missing event date for all
  starting events, i.e. `ENU`, `BTH` and `IMG` (enumeration, birth and
  immigration); end of study is the latest non-missing event data for
  closing events (coded as `OBE`). \cr\cr
  \code{vaRecordTests} checks again for valid codes, agreement between event
  coding and verbal autopsy coding, as well as agreement within verbal
  autopsy variables (the same set as in \code{vaVariableTests}).
}
\references{
Osman Sankoh and Peter Byass. The INDEPTH Network: filling vital gaps in global epidemiology. Int J Epi 2012; 41:579-588, Tables 1 and 2.
}
\seealso{
\code{\link{var-check-tools}} \code{\link{readRawHDSS}}
         \code{\link{preprocHDSS}} \code{\link{coreVariableTests}}
}

