% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocHDSS.R
\name{standardize-HDSS-var}
\alias{extractDate}
\alias{extractVAcode}
\alias{standardize-HDSS-var}
\title{Standardize raw HDSS variable}
\usage{
extractDate(x)

extractVAcode(x)
}
\arguments{
\item{x}{the variable to be processed (date, cause of death)}
}
\value{
For \code{extractDate}, a vector of \code{Date}s;
  for \code{extractVAcode}, a factor variable with standardized VA codes
}
\description{
These functions attempt to extract standardized data from specific variables
that may appear with variants in different HDSS sets
}
\details{
The function undertakes the following modifications:
\itemize{
    \item dates get stripped of trailing time stamps and converted to
          \code{Date}-variables, regardless of the separator used;
    \item \code{Sex} is coded as an f/m factor;
    \item for verbal autopsy variables, blanks are replaced with NAs;
    \item cause of death codes are matched against the canonical list of
          approved causes (based on WHO standard).
}
}
\note{
These functions will terminate with an error if their input is found to
  be too irregular for standardization
}
\seealso{
\code{\link{preprocHDSS}} \code{\link{INDEPTH_verbalAutopsy}}
}

