% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryHDSS.R
\name{summarize-HDSS}
\alias{overviewHDSS}
\alias{summarize-HDSS}
\alias{summarizeHDSS}
\title{Summarize a pre-processed HDSS data set}
\usage{
summarizeHDSS(x)

overviewHDSS(l)
}
\arguments{
\item{x}{The pre-processed HDSS data frame to be summarized}

\item{l}{a list of pre-processed HDSS data files}
}
\value{
For \code{summarizeHDSS} a named list of parameters, for
   \code{overviewHDSS} a data frame with centres/data files as rows and
   the parameters as columns.
}
\description{
Summarize key parameters of a pre-processed HDSS data set.
}
\details{
\code{summarizeHDSS} works for a single HDSS data set,
   \code{overviewHDSS} for a list of pre-processed data sets. Assuming that
   the list of HDSS files is based on raw data files following the INDEPTH
   naming conventions and read in by \code{readRawHDSS} before pre-processing,
   \code{overviewHDSS} tires to guess at the version of the data.
}
\seealso{
\code{\link{preprocHDSS}}
}

