% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryHDSS.R
\name{summary.HDSSdata}
\alias{summary.HDSSdata}
\title{Summary of pre-processed HDSS data}
\usage{
\method{summary}{HDSSdata}(x)
}
\arguments{
\item{x}{A pre-processed HDSS data file, as returned by
         \code{\link{preprocHDSS}}}
}
\value{
A data frame with key information as columns
}
\description{
Summary of pre-processed HDSS data
}

