% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varTests.R
\name{varTests-generic}
\alias{varTest_Exists}
\alias{varTest_ExistsNoMissing}
\alias{varTest_IsConstant}
\alias{varTest_IsSequential}
\alias{varTest_NoDuplicate}
\alias{varTest_NoMissing}
\alias{varTests-generic}
\title{Tools for checking variables in a data frame}
\usage{
varTest_Exists(x, v)

varTest_NoMissing(x, v)

varTest_ExistsNoMissing(x, v)

varTest_IsConstant(x, v)

varTest_NoDuplicate(x, v)

varTest_IsSequential(x, v)
}
\arguments{
\item{x}{the data frame where to check}

\item{v}{a character vector giving full variable names}
}
\value{
A data frame with three columns and as many rows as there are
  variable names specified. The columns are \code{Test}, which gives a short
  description of the test, \code{Variable}, which names the variable for
  which the test was applied, and \code{Pass}, which is a logical flag
  indicating whether or not the variable passed the test
}
\description{
These functions check for existence and properties of a specified variable
name in a given data frame.
}
\details{
These are simple tools suitable for checking whether a required
   variable is present in the data, and also whether there is a working
   primary key. The idea is that a data frame that passes these tests should
   be suitable for preprocessing. \cr\cr
   These tests are not suitable for checking the content of the variables,
   with the exception of checking for valid keys or crude missingness.\cr\cr
   Note that the data frame to be tested is the first argument, to make
   processing via the pipe operator easier.
}
\seealso{
\code{\link{dplyr::'\%>\%'}} \code{\link{recTests-generic}}
}

