% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varTests-HDSS.R
\name{varTests-HDSS}
\alias{coreVariableTests}
\alias{vaVariableTests}
\alias{varTests-HDSS}
\title{Check raw HDSS data}
\usage{
coreVariableTests(x)

vaVariableTests(x)
}
\arguments{
\item{x}{The data frame to be validated}
}
\value{
Each test returns a list with two entries:
  \itemize{
    \item{\code{flag}}{ a logical flag hat indicates whether all tests were passsed or not}
    \item{\code{text}}{ a data frame with one column that contains the error messages from the failed tests; empty if all tests were passed}
  }
}
\description{
These functions test whether a given dataframe read by \code{readRawHDSS} is
a valid HDSS data set in terms of variables present and a valid unique
identifier.
}
\details{
These functions provide a pragmatic way of testing core requirements
  for valid HDSS data. While is not a full validation of a formal HDSS
  specification, it is a set of reasonable tests: a data set that fails them
  will likely need manual intervention before further processing; a data set
  that passes them can be pre-processed before checking the actual content
  of the variables (see e.g. \code{coreRecTests}). \cr\cr
  \code{coreVariableTests} checks for the presence of the basic variables
  specified in Sankoh & Byass (2012), the presence of a valid unique record
  identifier, and for some crucial variables also the presence of missing
  values. \cr\cr
  \code{vaVariableTests} checks for the presence of verbal autopsy
  variables (given as \code{Cause1} to \code{Cause3} and \code{Likelihood1}
  to \code{Likelihood3}. This is not part of the core HDSS specification, and
  may or may not be present in a data set. \cr\cr
}
\references{
Osman Sankoh and Peter Byass. The INDEPTH Network: filling vital gaps in global epidemiology. Int J Epi 2012; 41:579-588, Tables 1 and 2.
}
\seealso{
\code{\link{var-check-tools}} \code{\link{readRawHDSS}}
         \code{\link{preprocHDSS}} \code{\link{coreRecTests}}
}

